+++ src/styles/components/_cardslider.scss
@@ -2,6 +2,8 @@ $color-circles: rgb(226, 245, 252);
 $color-waves: $color7;
 $top-wave: $color0;
 $bottom-wave: $color7;
+$contrastcard: $color2;
+$btncolor: $color4;
 
 .card-slider{
   display: flex;
@@ -36,6 +38,8 @@ $bottom-wave: $color7;
         max-width: 1200px;
         width: 75%;
         position: relative;
+      
+      
 
         ul{
           padding: 17px 0;
@@ -79,6 +83,8 @@ $bottom-wave: $color7;
           scrollbar-width: none;
 
           .card {
+            display: flex;
+            justify-content: flex-end;
             scroll-snap-align: start;
             height: 300px;
             list-style: none;
@@ -98,20 +104,65 @@ $bottom-wave: $color7;
                 border: 4px solid #fff;
               }
             }
+            .info-content{
+                height: 50%;
+                display: flex;
+    flex-direction: column;
+    justify-content: space-between;
+    align-items: center;
+    padding: 10px 14px;
+            }
 
             h2 {
-              font-weight: 500;
-              font-size: 1.56rem;
-              margin: 30px 0 5px;
+                color: #6A6D78;
+                font-size: 1.31rem;
+         
+             
             }
 
             span {
-              color: #6A6D78;
-              font-size: 1.31rem;
+                font-weight: 500;
+                font-size: 1.56rem;
             }
 
             &:hover {
               transform: scale(1.02); 
+            }
+            .footer {
+                max-height: 20px;
+                width: 100%;
+        
+                font-size: 0.85rem;
+                color: $contrastcard;
+                display: flex;
+                justify-content: space-between;
+                align-items: center;
+        
+                .more-btn {
+                  position: relative;
+                  border-radius: 50px;
+                  height: 30px;
+                  width: 30px;
+                  font-size: 1.7rem;
+                  color: $contrastcard;
+                  background: $btncolor;
+                  border: none;
+                  transition: transform 0.6s ease;
+        
+                  &:before {
+                    content: "+";
+                    position: absolute;
+                    top: 50%;
+                    left: 50%;
+                    transform: translate(-50%, -50%) translateY(-2px);
+                  }
+        
+                  &:hover {
+                    background: $contrast;
+                    color: $color-up;
+                    cursor: pointer;
+                  }
+                }
             }
           }
         }
