+++ src/styles/pages/_home.scss
@@ -45,11 +45,7 @@ header {
   
         .typing-text {
           font-weight: bold;
-         span{
-            content: " ";
-            height: 2rem;
-            display: block;
-         }
+         
         }
       }
